refactor(subgraph): consolidate helper methods
