+++ apps/subgraph/subgraph.yaml
@@ -59,13 +59,16 @@ dataSources:
         - name: FilecoinWarmStorageService
           file: ./abis/FilecoinWarmStorageService.json
       eventHandlers:
-        - event: DataSetCreated(indexed uint256,indexed uint256,uint256,uint256,uint256,address,address,address,string[],string[])
+        - event: DataSetCreated(indexed uint256,indexed
+            uint256,uint256,uint256,uint256,address,address,address,string[],string[])
           handler: handleFwssDataSetCreated
         - event: PieceAdded(indexed uint256,indexed uint256,(bytes),string[],string[])
           handler: handleFwssPieceAdded
-        - event: ServiceTerminated(indexed address,indexed  uint256,uint256,uint256,uint256)
+        - event: ServiceTerminated(indexed address,indexed
+            uint256,uint256,uint256,uint256)
           handler: handleFwssServiceTerminated
         - event: PDPPaymentTerminated(indexed uint256,uint256,uint256)
           handler: handleFwssPdpPaymentTerminated
-        - event: DataSetServiceProviderChanged(indexed uint256,indexed address,indexed address)
+        - event: DataSetServiceProviderChanged(indexed uint256,indexed address,indexed
+            address)
           handler: handleFwssDataSetServiceProviderChanged
